compartmentalized the user input form
