+++ src/components/Assets/Assets.css
@@ -1,7 +1,8 @@
 .assets_viewer {
     background-color: #E4E4E4; 
     width: 95%;
-    height: 85%;
+    height: 95%;
+    overflow: scroll;
 }
 
 .asset_tabs_container {
