+++ src/components/Assets/Assets.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import UserInputForm from '../UserInputForm/UserInputForm';
 import Categories from '../Categories/Categories';
