+++ src/components/UserInputForm/UserInputForm.css
@@ -1,7 +1,3 @@
-.userinputform {
-    display: flex;
-}
-
 .modal_popout_container {
     position: absolute;
     background: white;
